clean up rl implementation
