+++ recipes/configs/dev/3B_sft_for_grpo.yaml
@@ -28,7 +28,7 @@ tokenizer:
 
 # Dataset
 dataset:
-  _component_: torchtune.dev.grpo.gsm8k.gsm8k_sft
+  _component_: torchtune.rl.gsm8k.gsm8k_sft
   partition: 0-0/10
 seed: null
 shuffle: True
