+++ recipes/configs/qwen2_5/3B_qlora_grpo.yaml
@@ -36,7 +36,7 @@ tokenizer:
 
 # Dataset
 dataset:
-  _component_: torchtune.dev.grpo.gsm8k.gsm8k_dataset
+  _component_: torchtune.rl.gsm8k.gsm8k_dataset
   partition: 1-9/10
 seed: null
 shuffle: False
@@ -103,7 +103,7 @@ lr_scheduler:
   _component_: torchtune.training.lr_schedulers.get_cosine_schedule_with_warmup
   num_warmup_steps: 50
 loss:
-  _component_: torchtune.dev.grpo.loss.GRPOSimpleLoss
+  _component_: torchtune.rl.loss.GRPOSimpleLoss
   kl_coeff: 0.01
   epsilon: 0.2
 
@@ -118,11 +118,11 @@ compile: False  # pytorch compile, set to true for better perf/memory
 dtype: bf16
 
 reward_functions:
-    - _component_: torchtune.dev.rl.rewards.FormattedMathCorrectnessReward
+    - _component_: torchtune.rl.rewards.FormattedMathCorrectnessReward
       answer_tag: answer
       positive_reward: 10.0
       negative_reward: 0.0
-    - _component_: torchtune.dev.rl.rewards.ThinkingAnswerFormattingReward
+    - _component_: torchtune.rl.rewards.ThinkingAnswerFormattingReward
       think_tag: think
       answer_tag: answer
       positive_reward: 1.0
