+++ recipes/full_grpo_distributed.py
@@ -20,9 +20,9 @@ from torchdata.stateful_dataloader.sampler import StatefulDistributedSampler
 from torchtune import config, generation, modules, rlhf, training, utils
 from torchtune.config._utils import _get_component_from_path
 from torchtune.datasets import ConcatDataset
-from torchtune.dev.rl.generation import generate
-from torchtune.dev.rl.rewards import batched_rewards
-from torchtune.dev.rl.types import GRPOStats, GRPOTrajectory
+from torchtune.rl.generation import generate
+from torchtune.rl.rewards import batched_rewards
+from torchtune.rl.types import GRPOStats, GRPOTrajectory
 from torchtune.modules import local_kv_cache
 from torchtune.recipe_interfaces import FTRecipeInterface
 from torchtune.training import disable_dropout, DummyProfiler, PROFILER_KEY
@@ -190,7 +190,7 @@ class GRPOFullFinetuneRecipeDistributed(FTRecipeInterface):
         # sampler and dataloader depend on the tokenizer and loss_fn and should be
         # setup after both of these are initialized
         collate_name = cfg.get(
-            "collate_fn", "torchtune.dev.grpo.data.padded_collate_rl"
+            "collate_fn", "torchtune.rl.data.padded_collate_rl"
         )
         self._dataloader = self._setup_data(
             cfg_dataset=cfg.dataset,
