--- torchtune/dev/README.md
@@ -1,6 +1,0 @@
-## Torchtune dev components
-
-This directory houses experimental components.
-The purpose of `torchtune/dev` is to support bleeding-edge APIs
-with less stringent requirements on testing, documentation, or stability.
-The APIs in this folder are not public and are not guaranteed to adhere to backwards compatibility.
