+++ torchtune/rl/datatypes/vllm_completion_output.py
